+++ playgrounds/demo/app/pages/index.vue
@@ -1,27 +1,10 @@
 <script setup lang="ts">
 import usePageMeta from '~/composables/usePageMeta'
-import { useDashboard } from '@bitrix24/b24ui-nuxt/utils/dashboard'
-import { sleepAction } from '../utils/sleep'
 
 usePageMeta.setPageTitle('Bitrix24 UI - Demo')
-
-const { sidebarLoading, toggleLoading } = useDashboard({ sidebarLoading: ref(false), toggleLoading: () => {} })
-
-const makeToggleLoading = async (timeout: number = 1000) => {
-  toggleLoading?.(!sidebarLoading?.value)
-  await sleepAction(timeout)
-  toggleLoading?.(!sidebarLoading?.value)
-}
 </script>
 
 <template>
-  <B24Button
-    label="Test Reload"
-    loading-auto
-    class="mb-4"
-    @click="makeToggleLoading(1500)"
-  />
-
   <template
     v-for="(group) in usePageMeta.groups"
     :key="group.id"