+++ src/runtime/components/Sidebar.vue
@@ -14,9 +14,6 @@ export interface SidebarProps {
   class?: any
   b24ui?: Sidebar['slots']
 }
-/**
- * @todo add Pick<Xxxx
- */
 
 export interface SidebarSlots {
   default(props?: {}): any